Use an absolute install name for the shared library on Mac OS X.
Apple's Mac OS X dynamic linker uses an application's dependent
libraries' install names to find them at runtime. The install name
is essentially a pathname to the shared library that is record into
the application by the static linker. Since MySQL is typically
installed to /usr/local/mysql on Mac OS X, the dynamic linker won't
find the mysqlclient shared library as the aforementioned directory
is not one of the default locations where the dynamic linker looks
for libraries.

Hence, to allow the dynamic linker to locate the mysqlcient client
shared library, change the install name from just the library
filename to the absolute path to the library. This is accomplished
by setting the INSTALL_NAME_DIR property of the mysqlclient target,
and which specifies the directory portion of the install name field
of shared libraries on Mac OS X.