Print ISO date and PID with all error log messages.
The problem is that the currently available fields that compromise
the header of error log messages, which are the date and time, plus
the severity of the message, do not provide the enough information to
identify discontinuity in reporting or which process generated a
message.

To make such identification possible, this change includes the
process ID in the header so that any changes in its value will
identify the source (process) of the message and also indicate any
discontinuity in reporting.

Additionally, the format of the date and time is changed to the ISO
format (YYYY-MM-DD hh:mm:ss). The previous format was non-standard
and difficult to read.